gcp_function/requirements.txt
pandas
gcsfs
google-cloud-bigquery
google-cloud-storage